Split settings into it's own file
